time

+++ csfml_engine/lib/elapsed_time.c
@@ -8,7 +8,7 @@
 #include "my.h"
 #include "engine.h"
 
-bool elapsed_time_seconds(int seconds, float *increment, float *save_increment)
+bool elapsed_time_seconds(float seconds, float *increment, float *save_increment)
 {
     sfTime convert = sfClock_getElapsedTime(engine.win_settings.clock);
 
@@ -26,7 +26,7 @@ bool elapsed_time_seconds(int seconds, float *increment, float *save_increment)
     return false;
 }
 
-bool elapsed_time_milliseconds(int milliseconds, float *increment, float *save_increment)
+bool elapsed_time_milliseconds(float milliseconds, float *increment, float *save_increment)
 {
     sfTime convert = sfClock_getElapsedTime(engine.win_settings.clock);
 
@@ -44,7 +44,7 @@ bool elapsed_time_milliseconds(int milliseconds, float *increment, float *save_i
     return false;
 }
 
-bool elapsed_time_microseconds(int microseconds, float *increment, float *save_increment)
+bool elapsed_time_microseconds(float microseconds, float *increment, float *save_increment)
 {
     sfTime convert = sfClock_getElapsedTime(engine.win_settings.clock);
